+++ src/main/java/org/opensearch/knn/index/engine/KNNEngine.java
@@ -9,7 +9,6 @@ import com.google.common.collect.ImmutableSet;
 import org.opensearch.Version;
 import org.opensearch.common.ValidationException;
 import org.opensearch.knn.index.SpaceType;
-import org.opensearch.knn.index.VectorDataType;
 import org.opensearch.knn.index.engine.faiss.Faiss;
 import org.opensearch.knn.index.engine.lucene.Lucene;
 import org.opensearch.knn.index.engine.nmslib.Nmslib;
@@ -247,12 +246,12 @@ public enum KNNEngine implements KNNLibrary {
     }
 
     @Override
-    public boolean supportsRemoteIndexBuild(MethodComponentContext methodComponentContext, VectorDataType vectorDataType) {
-        return knnLibrary.supportsRemoteIndexBuild(methodComponentContext, vectorDataType);
+    public boolean supportsRemoteIndexBuild(KNNLibraryIndexingContext knnLibraryIndexingContext) {
+        return knnLibrary.supportsRemoteIndexBuild(knnLibraryIndexingContext);
     }
 
     @Override
-    public RemoteIndexParameters createRemoteIndexingParameters(KNNMethodContext knnMethodContext) {
-        return knnLibrary.createRemoteIndexingParameters(knnMethodContext);
+    public RemoteIndexParameters createRemoteIndexingParameters(Map<String, Object> parameters) {
+        return knnLibrary.createRemoteIndexingParameters(parameters);
     }
 }