+++ src/main/java/org/opensearch/knn/index/engine/faiss/Faiss.java
@@ -8,11 +8,10 @@ package org.opensearch.knn.index.engine.faiss;
 import com.google.common.collect.ImmutableMap;
 import org.opensearch.knn.common.KNNConstants;
 import org.opensearch.knn.index.SpaceType;
-import org.opensearch.knn.index.VectorDataType;
+import org.opensearch.knn.index.engine.KNNLibraryIndexingContext;
 import org.opensearch.knn.index.engine.KNNMethod;
 import org.opensearch.knn.index.engine.KNNMethodConfigContext;
 import org.opensearch.knn.index.engine.KNNMethodContext;
-import org.opensearch.knn.index.engine.MethodComponentContext;
 import org.opensearch.knn.index.engine.MethodResolver;
 import org.opensearch.knn.index.engine.NativeLibrary;
 import org.opensearch.knn.index.engine.ResolvedMethodContext;
@@ -23,6 +22,7 @@ import java.util.function.Function;
 
 import static org.opensearch.knn.common.KNNConstants.METHOD_HNSW;
 import static org.opensearch.knn.common.KNNConstants.METHOD_IVF;
+import static org.opensearch.knn.common.KNNConstants.NAME;
 
 /**
  * Implements NativeLibrary for the faiss native library
@@ -126,19 +126,21 @@ public class Faiss extends NativeLibrary {
      * Use the method name to route the check to the specific method class
      */
     @Override
-    public boolean supportsRemoteIndexBuild(MethodComponentContext methodComponentContext, VectorDataType vectorDataType) {
-        if (METHOD_HNSW.equals(methodComponentContext.getName())) {
-            if (methodComponentContext.getParameters() != null) {
-                return FaissHNSWMethod.supportsRemoteIndexBuild(methodComponentContext.getParameters(), vectorDataType);
+    public boolean supportsRemoteIndexBuild(KNNLibraryIndexingContext knnLibraryIndexingContext) {
+        if (knnLibraryIndexingContext != null) {
+            Map<String, Object> parameters = knnLibraryIndexingContext.getLibraryParameters();
+            if (METHOD_HNSW.equals(parameters.get(NAME))) {
+                return FaissHNSWMethod.supportsRemoteIndexBuild(parameters);
             }
         }
         return false;
     }
 
     @Override
-    public RemoteIndexParameters createRemoteIndexingParameters(KNNMethodContext knnMethodContext) {
-        if (METHOD_HNSW.equals(knnMethodContext.getMethodComponentContext().getName())) {
-            return FaissHNSWMethod.createRemoteIndexingParameters(knnMethodContext);
+    public RemoteIndexParameters createRemoteIndexingParameters(Map<String, Object> parameters) {
+        if (METHOD_HNSW.equals(parameters.get(NAME))) {
+            FaissHNSWMethod method = (FaissHNSWMethod) METHODS.get(METHOD_HNSW);
+            return method.createRemoteIndexingParameters(parameters);
         }
         throw new IllegalArgumentException("Unsupported method for remote indexing");
     }